Edit sales file
